requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
easyocr==1.7.1
filelock==3.12.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.285
langsmith==0.0.35
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pyclipper==1.3.0.post5
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.20
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
tifffile==2023.8.30
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
